Slowly fleshing out base tech concept.
Current problem is deepcopy(Technoloy(...)) is producing mulitple base, and breaking the Tech requirements system.
